fix(cmd/install): use separate chisel cache per worker
Use independent cache directory for chisel in each worker.
The reason is tricky to detect. When creating files in cache, Chisel
temporary saves a file as "<digest>.tmp" in the cache directory.[^1]
It later renames the file to "<digest>" within the same directory.[^2]
This is OK, if two chisel operations are sequential and/or do not try to
"create" the cache at the same time. But on concurrent operations, the
second process fails to rename the temporary file and eventually fails.

    error: cannot fetch from archive: rename
    /home/runner/.cache/chisel/sha256/6b8cc68643d18250ab297c4f6d427a8778b1d1534e1a3033fff0f221fa20a419.tmp
    /home/runner/.cache/chisel/sha256/6b8cc68643d18250ab297c4f6d427a8778b1d1534e1a3033fff0f221fa20a419:
    no such file or directory

[^1]: https://github.com/canonical/chisel/blob/main/internal/cache/cache.go#L112
[^2]: https://github.com/canonical/chisel/blob/main/internal/cache/cache.go#L80